+++ src/imio/helpers/tests/test_transmogrifier.py
@@ -55,6 +55,10 @@ class TestTesting(IntegrationTestCase):
         self.assertListEqual([1, 2, 3], sorted(filter_keys(dic, ['a', 'b', 'c']).values()))
         self.assertListEqual(['a', 'c'], sorted(filter_keys(dic, ['a', 'c']).keys()))
         self.assertListEqual([1, 3], sorted(filter_keys(dic, ['a', 'c']).values()))
+        del dic['c']
+        self.assertListEqual(['a', 'b', 'c'], sorted(filter_keys(dic, ['a', 'b', 'c']).keys()))
+        self.assertListEqual([None, 1, 2], sorted(filter_keys(dic, ['a', 'b', 'c']).values()))
+        self.assertListEqual([1, 2, ''], sorted(filter_keys(dic, ['a', 'b', 'c'], unfound='').values()))
 
     def test_get_obj_from_path(self):
         folder = self.portal.folder